Refatorações e melhorias no código, e modificação da Entidade Pedido para Venda.

+++ src/main/java/com/jbmotos/services/impl/FornecedorServiceImpl.java
@@ -2,7 +2,6 @@ package com.jbmotos.services.impl;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,6 +23,8 @@ import com.jbmotos.services.exception.RegraDeNegocioException;
 @Service
 public class FornecedorServiceImpl implements FornecedorService {
 
+	private final String FORNECEDOR_NAO_ENCONTRADO = "Fornecedor não encrontrado para o CNPJ informado.";
+
     private final String ERRO_SALVAR_FORNECEDOR = "Erro ao tentar salvar Fornecedor";
 
     @Autowired
@@ -43,10 +44,8 @@ public class FornecedorServiceImpl implements FornecedorService {
 		fornecedor.setStatusFornecedor(Situacao.ATIVO);
 		fornecedor.setDataHoraCadastro(LocalDateTime.now());
 
-		Optional<Endereco> enderecoOptional = enderecoService.buscarEnderecoPorId(fornecedorDTO.getEndereco());
-		if (enderecoOptional.isPresent()) {
-			fornecedor.setEndereco(enderecoOptional.get());
-		}
+		Endereco endereco = enderecoService.buscarEnderecoPorId(fornecedorDTO.getEndereco());
+		fornecedor.setEndereco(endereco);
 
 		return fornecedorRepository.save(fornecedor);
 	}
@@ -59,9 +58,9 @@ public class FornecedorServiceImpl implements FornecedorService {
 
     @Override
     @Transactional(readOnly = true)
-    public Optional<Fornecedor> buscarFornecedorPorCNPJ(String cnpj) {
-        checarCnpjFornecedorExistente(cnpj);
-        return fornecedorRepository.findFornecedorByCnpj(cnpj);
+    public Fornecedor buscarFornecedorPorCNPJ(String cnpj) {
+        return fornecedorRepository.findFornecedorByCnpj(cnpj)
+        		.orElseThrow(() -> new ObjetoNaoEncontradoException(FORNECEDOR_NAO_ENCONTRADO));
     }
 
     @Override
@@ -77,43 +76,38 @@ public class FornecedorServiceImpl implements FornecedorService {
 	@Override
 	@Transactional
 	public Situacao alternarStatusFornecedor(String cnpj) {
-		Optional<Fornecedor> fornecedorOptional = buscarFornecedorPorCNPJ(cnpj);
-		if (fornecedorOptional.isPresent()) {
-			if (fornecedorOptional.get().getStatusFornecedor().equals(Situacao.ATIVO)) {
-				fornecedorOptional.get().setStatusFornecedor(Situacao.INATIVO);
-			} else if (fornecedorOptional.get().getStatusFornecedor().equals(Situacao.INATIVO)) {
-				fornecedorOptional.get().setStatusFornecedor(Situacao.ATIVO);
-			}
-			fornecedorRepository.save(fornecedorOptional.get());
-			return fornecedorOptional.get().getStatusFornecedor();
+		Fornecedor fornecedor = buscarFornecedorPorCNPJ(cnpj);
+		if (fornecedor.getStatusFornecedor().equals(Situacao.ATIVO)) {
+			fornecedor.setStatusFornecedor(Situacao.INATIVO);
+		} else if (fornecedor.getStatusFornecedor().equals(Situacao.INATIVO)) {
+			fornecedor.setStatusFornecedor(Situacao.ATIVO);
 		}
-		return null;
+		fornecedorRepository.save(fornecedor);
+		return fornecedor.getStatusFornecedor();
 	}
 
 	@Override
 	@Transactional
 	public Fornecedor atualizarFornecedor(FornecedorDTO fornecedorDTO) {
-		Optional<Fornecedor> fornecedorOptional = buscarFornecedorPorCNPJ(fornecedorDTO.getCnpj());
-		if (fornecedorOptional.isPresent()) {
-			LocalDateTime dateTime = fornecedorOptional.get().getDataHoraCadastro();
-			Fornecedor fornecedor = mapper.map(fornecedorDTO, Fornecedor.class);
-			fornecedor.setDataHoraCadastro(dateTime);
+		Fornecedor fornecedor = mapper.map(fornecedorDTO, Fornecedor.class);
 
-			Optional<Endereco> enderecoOptional = enderecoService.buscarEnderecoPorId(fornecedorDTO.getEndereco());
-			if (enderecoOptional.isPresent()) {
-				fornecedor.setEndereco(enderecoOptional.get());
-			}
-			return fornecedorRepository.save(fornecedor);
-		}
-		return null;
+		LocalDateTime dateTime = buscarFornecedorPorCNPJ(fornecedorDTO.getCnpj()).getDataHoraCadastro();
+		fornecedor.setDataHoraCadastro(dateTime);
+
+		Endereco endereco = enderecoService.buscarEnderecoPorId(fornecedorDTO.getEndereco());
+		fornecedor.setEndereco(endereco);
+
+		return fornecedorRepository.save(fornecedor);
 	}
 
-    @Override
-    @Transactional
-    public void deletarFornecedor(String cnpj) {
-        checarCnpjFornecedorExistente(cnpj);
-        fornecedorRepository.deleteFornecedorByCnpj(cnpj);
-    }
+	@Override
+	@Transactional
+	public void deletarFornecedor(String cnpj) {
+		if (!fornecedorRepository.existsFornecedorByCnpj(cnpj)) {
+			throw new ObjetoNaoEncontradoException(FORNECEDOR_NAO_ENCONTRADO);
+		}
+		fornecedorRepository.deleteFornecedorByCnpj(cnpj);
+	}
 
     @Override
     public void validarCnpjFornecedorParaSalvar(String cnpj) {
@@ -125,13 +119,6 @@ public class FornecedorServiceImpl implements FornecedorService {
     @Override
     public List<Fornecedor> filtrarFornecedoresPorCnpjDiferente(FornecedorDTO fornecedorDTO) {
         return fornecedorRepository.findByCnpjNot(fornecedorDTO.getCnpj());
-    }
-
-    @Override
-    public void checarCnpjFornecedorExistente(String cnpj) {
-        if (!fornecedorRepository.existsFornecedorByCnpj(cnpj)) {
-            throw new ObjetoNaoEncontradoException("Fornecedor não encrontrado para o CNPJ informado.");
-        }
     }
 
     @Override
